+++ test/e2e/configurer/chain/commands_btcstaking.go
@@ -19,7 +19,6 @@ import (
 	asig "github.com/babylonlabs-io/babylon/crypto/schnorr-adaptor-signature"
 	"github.com/babylonlabs-io/babylon/test/e2e/containers"
 	bbn "github.com/babylonlabs-io/babylon/types"
-	btcctypes "github.com/babylonlabs-io/babylon/x/btccheckpoint/types"
 	bstypes "github.com/babylonlabs-io/babylon/x/btcstaking/types"
 )
 
@@ -45,7 +44,8 @@ func (n *NodeConfig) CreateFinalityProvider(walletAddrOrName string, btcPK *bbn.
 func (n *NodeConfig) CreateBTCDelegation(
 	btcPk *bbn.BIP340PubKey,
 	pop *bstypes.ProofOfPossessionBTC,
-	stakingTxInfo *btcctypes.TransactionInfo,
+	stakingTx []byte,
+	inclusionProof *bstypes.InclusionProof,
 	fpPK *bbn.BIP340PubKey,
 	stakingTimeBlocks uint16,
 	stakingValue btcutil.Amount,
@@ -69,8 +69,15 @@ func (n *NodeConfig) CreateBTCDelegation(
 	require.NoError(n.t, err)
 
 	// get staking tx info hex
-	stakingTxInfoHex, err := stakingTxInfo.ToHexStr()
-	require.NoError(n.t, err)
+	stakingTxHex := hex.EncodeToString(stakingTx)
+
+	// get inclusion proof hex
+	var inclusionProofHex string
+
+	if inclusionProof != nil {
+		inclusionProofHex, err = inclusionProof.MarshalHex()
+		require.NoError(n.t, err)
+	}
 
 	fpPKHex := fpPK.MarshalHex()
 
@@ -93,14 +100,20 @@ func (n *NodeConfig) CreateBTCDelegation(
 
 	cmd := []string{
 		"babylond", "tx", "btcstaking", "create-btc-delegation",
-		btcPkHex, popHex, stakingTxInfoHex, fpPKHex, stakingTimeString, stakingValueString, slashingTxHex, delegatorSigHex, unbondingTxHex, unbondingSlashingTxHex, unbondingTimeStr, unbondingValueStr, delUnbondingSlashingSigHex,
+		btcPkHex, popHex, stakingTxHex, inclusionProofHex, fpPKHex, stakingTimeString, stakingValueString, slashingTxHex, delegatorSigHex, unbondingTxHex, unbondingSlashingTxHex, unbondingTimeStr, unbondingValueStr, delUnbondingSlashingSigHex,
 		fmt.Sprintf("--from=%s", fromWalletName), containers.FlagHome, flagKeyringTest,
 		n.FlagChainID(), "--log_format=json",
 	}
 
+	// gas price
+	cmd = append(cmd, "--gas-prices=0.002ubbn")
+
 	if generateOnly {
 		cmd = append(cmd, "--generate-only")
 	} else {
+		// gas
+		cmd = append(cmd, "--gas=auto", "--gas-adjustment=1.3")
+		// broadcast stuff
 		cmd = append(cmd, "-b=sync", "--yes")
 	}
 
@@ -137,7 +150,7 @@ func (n *NodeConfig) AddCovenantSigs(covPK *bbn.BIP340PubKey, stakingTxHash stri
 	// used key
 	cmd = append(cmd, "--from=val")
 	// gas
-	cmd = append(cmd, "--gas=auto", "--gas-prices=1ubbn", "--gas-adjustment=1.3")
+	cmd = append(cmd, "--gas=auto", "--gas-adjustment=1.3")
 
 	_, _, err := n.containerManager.ExecTxCmd(n.t, n.chainId, n.Name, cmd)
 	require.NoError(n.t, err)
@@ -227,4 +240,17 @@ func (n *NodeConfig) BTCUndelegate(stakingTxHash *chainhash.Hash, delUnbondingSi
 	_, _, err := n.containerManager.ExecTxCmd(n.t, n.chainId, n.Name, cmd)
 	require.NoError(n.t, err)
 	n.LogActionF("successfully added signature on unbonding tx from delegator")
+}
+
+func (n *NodeConfig) AddBTCDelegationInclusionProof(
+	stakingTxHash *chainhash.Hash,
+	inclusionProof *bstypes.InclusionProof) {
+	n.LogActionF("activate delegation by adding inclusion proof")
+	inclusionProofHex, err := inclusionProof.MarshalHex()
+	require.NoError(n.t, err)
+
+	cmd := []string{"babylond", "tx", "btcstaking", "add-btc-inclusion-proof", stakingTxHash.String(), inclusionProofHex, "--from=val"}
+	_, _, err = n.containerManager.ExecTxCmd(n.t, n.chainId, n.Name, cmd)
+	require.NoError(n.t, err)
+	n.LogActionF("successfully added inclusion proof")
 }