+++ test/e2e/e2e_test.go
@@ -23,6 +23,9 @@ func TestBTCTimestampingTestSuite(t *testing.T) {
 func TestBTCStakingTestSuite(t *testing.T) {
 	suite.Run(t, new(BTCStakingTestSuite))
 }
+func TestBTCStakingPreApprovalTestSuite(t *testing.T) {
+	suite.Run(t, new(BTCStakingPreApprovalTestSuite))
+}
 
 // TestSoftwareUpgradeV1TestnetTestSuite tests software upgrade of v1 testnet end-to-end
 func TestSoftwareUpgradeV1TestnetTestSuite(t *testing.T) {