+++ x/epoching/keeper/drop_validator_msg_decorator_test.go
@@ -24,7 +24,7 @@ func TestDropValidatorMsgDecorator(t *testing.T) {
 		{&stakingtypes.MsgEditValidator{}, false},
 	}
 
-	decorator := NewDropValidatorMsgDecorator(Keeper{})
+	decorator := NewDropValidatorMsgDecorator(&Keeper{})
 
 	for _, tc := range testCases {
 		res := decorator.IsValidatorRelatedMsg(tc.msg)